- Redirecting stderr to a file no longer has the effect of showing error   messages and assertion failure notifications as requesters. The exception   is in redirecting stderr to NIL: which will prompt the requester use.
- gettimeofday() now calls GetSysTime() rather than DateStamp() to obtain
  the current system time. This resolves granularity issues since the
  DateStamp() result was only accurate by 1/50 of a second.

- The "ptrdiff_t" definition in <stddef.h> now defaults to type 'int' rather
  than 'long int'.


git-svn-id: file:///Users/olsen/Code/migration-svn-zu-git/logical-line-staging/clib2/trunk@14791 87f5fb63-7c3d-0410-a384-fd976d0f7a62

+++ library/include/stddef.h
@@ -1,5 +1,5 @@
 /*
- * $Id: stddef.h,v 1.2 2005-01-02 09:07:21 obarthel Exp $
+ * $Id: stddef.h,v 1.3 2005-01-08 10:21:28 obarthel Exp $
  *
  * :ts=4
  *
@@ -52,7 +52,7 @@ extern "C" {
 
 /****************************************************************************/
 
-typedef long int ptrdiff_t;
+typedef int ptrdiff_t;
 typedef unsigned int size_t;
 typedef unsigned int wchar_t;
